Make box input dynamic
Updates display count on input rather than on change so that enter or
touch-out is not necessary when changing number in display count box.

Once slider reaches 100, disply count box is always visible. This way
there is no issue with debouncing when editing the count in the box. For
example, erasing the last digit from 100 to attempt to set to 101 would
cause the box to disappear if the box was meant to be hidden for values
less than 101.